#100 : Instead of pixel sampling, you can call the with_font_bg function in the initializer to specify an explicit background color to have converted to black/transparent.

+++ bracket-terminal/src/hal/curses/main_loop.rs
@@ -134,7 +134,7 @@ pub fn main_loop<GS: GameState>(mut bterm: BTerm, mut gamestate: GS) -> Result<(
                         window.mvaddch(
                             st.height as i32 - (y as i32 + 1),
                             x as i32,
-                            to_char(t.glyph),
+                            to_char(t.glyph as u8),
                         );
                         idx += 1;
                     }
@@ -150,7 +150,7 @@ pub fn main_loop<GS: GameState>(mut bterm: BTerm, mut gamestate: GS) -> Result<(
                     window.mvaddch(
                         st.height as i32 - (y as i32 + 1),
                         x as i32,
-                        to_char(t.glyph),
+                        to_char(t.glyph as u8),
                     );
                 }
             }
